Move the examples into their own unpublished module
Following review of the first attempt, which parked them in the
sanitizer's test sources.  That kept them out of the jar but cost more
than it should have: test sources compile against the test classpath, so
an example could quietly start importing commons-codec or a
package-private test helper and still build, and CodeQL's autobuilder
skips test compilation, so three security-relevant sample policies that
people copy verbatim dropped out of the scan.

An examples module keeps them as main sources and gets the same result.
It declares maven.deploy.skip, so nothing is published, and its only
compile dependency is the sanitizer itself, which turns the "examples
use nothing but the public API" convention into something the build
enforces.  Verified: adding an import of commons-codec or htmlparser to
UrlTextExample now fails compilation.  java8-shim is runtime scope --
the released jar shades it in, so a copied example needs nothing extra,
but it stays off the compile classpath.

ExamplesTest and UrlTextExampleTest move with the code they cover and
into org.owasp.html.examples, which also stops them sharing a package
with the library.  They only ever used the examples' public run methods,
so nothing had to change but the package line and two now-redundant
imports.  Test totals are unchanged: 430 in the sanitizer plus 7 here,
against 437 before.

The published jar carries 121 classes under org/owasp/html and
org/owasp/shim only, with nothing under examples.
