+++ examples/pom.xml
@@ -1,0 +1,49 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <artifactId>examples</artifactId>
+  <version>20000101-SNAPSHOT</version>
+  <packaging>jar</packaging>
+  <parent>
+    <groupId>com.googlecode.owasp-java-html-sanitizer</groupId>
+    <artifactId>parent</artifactId>
+    <version>20000101-SNAPSHOT</version>
+  </parent>
+
+  <name>examples</name>
+
+  <properties>
+    <!-- Sample code for people to read and copy; never publish it.  See #180. -->
+    <maven.deploy.skip>true</maven.deploy.skip>
+  </properties>
+  <url>https://github.com/OWASP/java-html-sanitizer</url>
+  <description>
+    Worked examples of building a sanitization policy.  These are meant to be
+    read and copied into your own project, not depended upon, so they are built
+    and tested here but deliberately not released.
+
+    The only compile dependency is the sanitizer itself, so an example that
+    reached for anything outside the library's public API would fail the build.
+  </description>
+
+  <dependencies>
+    <dependency>
+      <groupId>com.googlecode.owasp-java-html-sanitizer</groupId>
+      <artifactId>owasp-java-html-sanitizer</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <!-- The released jar shades the shim in, so a copied example needs nothing
+         extra.  Runtime scope keeps it off the compile classpath, so the
+         examples cannot accidentally reach past the public API. -->
+    <dependency>
+      <groupId>com.googlecode.owasp-java-html-sanitizer</groupId>
+      <artifactId>java8-shim</artifactId>
+      <version>${project.version}</version>
+      <scope>runtime</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.junit.jupiter</groupId>
+      <artifactId>junit-jupiter</artifactId>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+</project>
