+++ examples/src/main/java/org/owasp/html/examples/EbayPolicyExample.java
