+++ examples/src/main/java/org/owasp/html/examples/SlashdotPolicyExample.java
