+++ examples/src/test/java/org/owasp/html/examples/ExamplesTest.java
@@ -25,15 +25,12 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-package org.owasp.html;
+package org.owasp.html.examples;
 
 import java.io.IOException;
 import java.io.StringReader;
 
 import org.junit.jupiter.api.Test;
-import org.owasp.html.examples.EbayPolicyExample;
-import org.owasp.html.examples.SlashdotPolicyExample;
-import org.owasp.html.examples.UrlTextExample;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
